chatbot_app/pyproject.toml
[project]
name = "langchain-chatbot"
version = "0.1.0"
description = "Basic LangChain chatbot with CLI and Streamlit interfaces"
authors = [{name = "vapmail16",email = "vapmail16@gmail.com"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "*"
langchain = "*"
streamlit = "*"
python-dotenv = "*"
langchain-community = "^0.3.20"
duckduckgo-search = "^7.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
